Update mainFunction.cpp
The major update of the software from the versions 1.15.* to 1.17.**

Introduced new rules and syntax for storing the website parameters. The speed is improved for rendering the website.

The database and file format are compatible with the version 1.**.**

However, the following two documents must be changed into new syntax: database entry mainTextInitializer, and the setup file st*secretCode*.txt inside the website root folder.