Update 1.17.17 mainFunction.cpp
The version number is updated. The update 1.17.17 provides an option of assigning trivial versions for the exam to students, and an option for creating an exam in which problems are not displayed. The no-display-problems is convenient for in-class exams in which the web support is focused only on providing auto-grader, but not on displaying formulations and solutions to students.