Update 2.0.0 to HBF_HdB.cpp
Changed the way in which the nodes of the B-tree are stored on the disk. The nodes are now stored as a vector. The raw files are less human readable. However, the files can now be read only once when passing through the tree. This has increased the speed significantly.